refactor: refactored moving assertCar to CarTest from RacingTest
