+++ src/test/java/racing/RacingTest.java
@@ -36,21 +36,4 @@ class RacingTest {
         });
         assertThat(sumAfterStartGame.get()).isGreaterThan(0);
     }
-
-    @Test
-    void assertCar() {
-        Car car = new Car();
-        int testValueLessThanFour = 3;
-        int testValueGreaterThanFour = 5;
-
-        car.moveCar();
-
-        assertThat(car.getDistance()).isEqualTo(1);
-
-        car.validateForMove(testValueLessThanFour);
-        assertThat(car.getDistance()).isEqualTo(1);
-
-        car.validateForMove(testValueGreaterThanFour);
-        assertThat(car.getDistance()).isEqualTo(2);
-    }
 }
